search accessible clone() to fix java17 compatability (#130)
XMLGregorianCalender is not cloneable with the current implementation
when using java 17. The implementation did not select the method
XMLGregorianCalendar.clone() which is accessible to anyone, but rather
the method XMLGregorianCalenderImpl.clone() to which access is
prohibited in recent java versions ending up in an IllegalAccessError.

This fixes the problem by searching for the first clone() method in
the classes hierarchy that is public and accessible. This removes
the need of calling setAccessible because the method will be accessible.
For java8 it will still use the old logic because Method.canAccess()
does not exist until java9.